Ben scheduler

+++ app/models/event.py
@@ -2,7 +2,7 @@ from database import db
 
 class Event(db.Model):
     id = db.Column(db.Integer, primary_key=True)
-    facebook_id = db.Column(db.Integer)
+    facebook_id = db.Column(db.BigInteger)
     
     send_timestamp = db.Column(db.DateTime)
     metadata_json = db.Column(db.Text)
@@ -11,7 +11,7 @@ class Event(db.Model):
     # 1 - morning info card
     # 2 - morning lyft
     # 3 - afternoon lyft
-
+ 
     def __init__(self, facebook_id):
         self.facebook_id = facebook_id
